requirements-dev.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.2
django-bootstrap-form==3.4
django-filter==21.1
djangorestframework==3.13.1
fonttools==4.29.1
idna==3.3
kiwisolver==1.3.2
loguru==0.6.0
matplotlib==3.5.1
numpy==1.22.2
packaging==21.3
Pillow==9.0.1
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
requests==2.27.1
response==0.5.0
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
urllib3==1.26.8
